use long

+++ src/main/java/org/spiget/resourcemanagerfetcher/SpigetRestFetcher.java
@@ -114,7 +114,7 @@ public class SpigetRestFetcher {
 
 		try {
 			databaseClient.updateStatus("fetch.rest.start", startTime);
-			int lastEnd = databaseClient.getStatus("fetch.rest.end", 0);
+			long lastEnd = databaseClient.getStatus("fetch.rest.end", 0);
 			databaseClient.updateStatus("fetch.rest.lastEnd", lastEnd);
 			databaseClient.updateStatus("fetch.rest.end", 0);
 		} catch (Exception e) {
